utils.parser: moved in lexemSimplifyForView
